requirements.txt
alembic==0.8.6
astroid==1.4.6
boto3==1.3.1
botocore==1.4.17
certifi==2016.2.28
click==6.6
colorama==0.3.7
colour==0.1.2
datadog==0.12.0
decorator==4.0.9
docutils==0.12
Flask==0.11.1
Flask-Cache==0.13.1
Flask-JWT==0.3.2
Flask-Login==0.3.2
Flask-Migrate==1.8.0
Flask-MySQLdb==0.2.0
Flask-Script==2.0.5
Flask-Session==0.2.3
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
infinity==1.4
intervals==0.7.1
itsdangerous==0.24
Jinja2==2.8
jmespath==0.9.0
lazy-object-proxy==1.2.2
Mako==1.0.4
MarkupSafe==0.23
mock==2.0.0
mysqlclient==1.3.7
nose==1.3.7
pbr==1.10.0
Pillow==3.2.0
PyJWT==1.4.0
pylint==1.5.6
python-dateutil==2.5.3
python-editor==1.0.1
redis==2.10.5
requests==2.9.1
selenium==2.53.2
simplejson==3.8.2
six==1.10.0
SQLAlchemy==1.0.13
urllib3==1.15.1
validators==0.10
Werkzeug==0.11.10
wrapt==1.10.8
WTForms==2.1
WTForms-Components==0.10.0
